refactor: 套用數個異動
1. 重構 LiveChatCatcher 函式庫的程式碼結構。
2. 將大部分與 YouTube 有關的方法都遷移到 LiveChatCatcher 函式庫內。
3. 更新 LiveChatCatcher 的 README.md 檔案。
4. 更新 YouTube 聊天室捕手的 README.md 檔案。
