+++ LiveChatCatcher/Extensions/JsonElementExtension.cs
@@ -1,5 +1,4 @@
 ﻿using System.Text.Json;
-using Rubujo.YouTube.Utility.Sets;
 
 namespace Rubujo.YouTube.Utility.Extensions;
