+++ LiveChatCatcher/LiveChatCatcher.ParseJson.cs
@@ -1,12 +1,4 @@
-﻿using AngleSharp.Dom;
-using AngleSharp.Html.Dom;
-using AngleSharp.Html.Parser;
-using System.Drawing;
-using System.Net;
-using System.Net.Http.Headers;
-using System.Security.Cryptography;
-using System.Text;
-using System.Text.Json;
+﻿using System.Text.Json;
 using Rubujo.YouTube.Utility.Extensions;
 using Rubujo.YouTube.Utility.Models;
 using Rubujo.YouTube.Utility.Sets;
@@ -14,410 +6,12 @@ using Rubujo.YouTube.Utility.Sets;
 namespace Rubujo.YouTube.Utility;
 
 /// <summary>
-/// LiveChatCatcher 的私人方法
+/// LiveChatCatcher 的解析 JSON 資料方法
 /// <para>參考 1：https://takikomiprogramming.hateblo.jp/entry/2020/07/21/114851</para>
 /// <para>參考 2：https://yasulab-pg.com/%E3%80%90python%E3%80%91youtube-live%E3%81%AE%E3%82%A2%E3%83%BC%E3%82%AB%E3%82%A4%E3%83%96%E3%81%8B%E3%82%89%E3%83%81%E3%83%A3%E3%83%83%E3%83%88%E3%82%92%E5%8F%96%E5%BE%97%E3%81%99%E3%82%8B/</para>
 /// </summary>
 public partial class LiveChatCatcher
 {
-    /// <summary>
-    /// 取得 ytcfg 資料
-    /// </summary>
-    /// <param name="videoID">字串，影片 ID</param>
-    /// <returns>InitialData</returns>
-    private InitialData GetYTConfigData(string videoID)
-    {
-        InitialData initialData = new()
-        {
-            YTConfigData = new()
-        };
-
-        string url = SharedIsStreaming ? $"{StringSet.Origin}/live_chat?v={videoID}" :
-            $"{StringSet.Origin}/watch?v={videoID}";
-
-        HttpRequestMessage httpRequestMessage = new(HttpMethod.Get, url);
-
-        if (!string.IsNullOrEmpty(SharedCookies))
-        {
-            SetHttpRequestMessageHeader(httpRequestMessage);
-        }
-
-        HttpResponseMessage? httpResponseMessage = SharedHttpClient?.SendAsync(httpRequestMessage)
-            .GetAwaiter()
-            .GetResult();
-
-        string? htmlContent = httpResponseMessage?.Content.ReadAsStringAsync()
-            .GetAwaiter()
-            .GetResult();
-
-        if (httpResponseMessage?.StatusCode == HttpStatusCode.OK)
-        {
-            if (string.IsNullOrEmpty(htmlContent))
-            {
-                RaiseOnLogOutput(
-                    EnumSet.LogType.Error,
-                    "[GetYTConfigData()] 發生錯誤，變數 \"htmlContent\" 為空白或是 null！");
-
-                return initialData;
-            }
-
-            HtmlParser htmlParser = new();
-            IHtmlDocument htmlDocument = htmlParser.ParseDocument(htmlContent);
-            IHtmlCollection<IElement> scriptElements = htmlDocument.QuerySelectorAll("script");
-            IElement elementYtCfg = scriptElements
-                .FirstOrDefault(n => n.InnerHtml.Contains("ytcfg.set({"))!;
-
-            if (elementYtCfg == null)
-            {
-                RaiseOnLogOutput(
-                    EnumSet.LogType.Error,
-                    "[GetYTConfigData()] 發生錯誤，變數 \"elementYtCfg\" 為 null！");
-
-                return initialData;
-            }
-
-            // TODO: 2023/6/13 考慮是否待修改。
-            // 可以參考 1：https://github.com/abhinavxd/youtube-live-chat-downloader/blob/v1.0.5/yt_chat.go#L140
-            // 可以參考 2：https://github.com/xenova/chat-downloader/blob/master/chat_downloader/sites/youtube.py#L443
-            string jsonYtCfg = elementYtCfg.InnerHtml;
-
-            if (SharedIsStreaming)
-            {
-                jsonYtCfg = jsonYtCfg.Replace("ytcfg.set(", string.Empty);
-
-                int endTokenIndex = jsonYtCfg.LastIndexOf("});");
-
-                // 要補回最後一個 "}"。
-                jsonYtCfg = jsonYtCfg[..(endTokenIndex + 1)];
-            }
-            else
-            {
-                int startTokenIndex = jsonYtCfg.IndexOf("ytcfg.set({"),
-                    endTokenIndex = jsonYtCfg.IndexOf("});");
-
-                // 只擷取要的部分。
-                jsonYtCfg = jsonYtCfg.Substring(startTokenIndex, endTokenIndex);
-                jsonYtCfg = jsonYtCfg.Replace("ytcfg.set(", string.Empty);
-
-                // 重新再在找一次。
-                endTokenIndex = jsonYtCfg.LastIndexOf("});");
-
-                // 要補回最後一個 "}"。
-                jsonYtCfg = jsonYtCfg[..(endTokenIndex + 1)];
-            }
-
-            JsonElement jeYtCfg = JsonSerializer.Deserialize<JsonElement>(jsonYtCfg);
-
-            initialData.YTConfigData = ParseYtCfg(jeYtCfg);
-
-            IElement elementYtInitialData = SharedIsStreaming ?
-                scriptElements.FirstOrDefault(n => n.InnerHtml.Contains("window[\"ytInitialData\"] ="))! :
-                scriptElements.FirstOrDefault(n => n.InnerHtml.Contains("var ytInitialData ="))!;
-
-            string jsonYtInitialData = SharedIsStreaming ?
-                elementYtInitialData.InnerHtml.Replace("window[\"ytInitialData\"] = ", string.Empty) :
-                elementYtInitialData.InnerHtml.Replace("var ytInitialData = ", string.Empty);
-
-            if (jsonYtInitialData.EndsWith(';'))
-            {
-                jsonYtInitialData = jsonYtInitialData[0..^1];
-            }
-
-            JsonElement jeYtInitialData = JsonSerializer.Deserialize<JsonElement>(jsonYtInitialData);
-
-            initialData.YTConfigData.Continuation = SharedIsStreaming ?
-                ParseFirstTimeContinuation(jeYtInitialData)[0] :
-                ParseReplayContinuation(jeYtInitialData);
-
-            // 若是直播中的影片時，剛載入頁面就影片聊天室的內容，這些資料也需要處理。
-            if (SharedIsStreaming)
-            {
-                initialData.Messages = ParseActions(jeYtInitialData);
-            }
-        }
-        else
-        {
-            string errorMessage = $"[{DateTime.Now}]（{typeof(HttpClient).Name}）：" +
-                $"連線發生錯誤，錯誤碼：{httpResponseMessage?.StatusCode} " +
-                $"{(httpResponseMessage != null ? $"({(int)(httpResponseMessage.StatusCode)})" : string.Empty)}{Environment.NewLine}" +
-                $"接收到的內容：{Environment.NewLine}" +
-                $"{htmlContent}{Environment.NewLine}";
-
-            RaiseOnLogOutput(
-                EnumSet.LogType.Error,
-                errorMessage);
-        }
-
-        return initialData;
-    }
-
-    /// <summary>
-    /// 取得 JsonElement
-    /// </summary>
-    /// <param name="ytConfigData">YTConfig</param>
-    /// <returns>JsonElement</returns>
-    private JsonElement GetJsonElement(YTConfigData ytConfigData)
-    {
-        JsonElement jsonElement = new();
-
-        string methodName = SharedIsStreaming ? "get_live_chat" : "get_live_chat_replay";
-        string url = $"{StringSet.Origin}/youtubei/v1/live_chat/{methodName}?key={ytConfigData.APIKey}";
-
-        // 當 ytConfigData.Continuation 為 null 或空值時，則表示已經抓取完成。
-        if (!string.IsNullOrEmpty(ytConfigData.Continuation))
-        {
-            // 當沒有時才指定，後續不更新。
-            if (string.IsNullOrEmpty(ytConfigData.InitPage))
-            {
-                string apiType = methodName.Replace("get_", string.Empty);
-
-                ytConfigData.InitPage = $"{StringSet.Origin}/{apiType}/?continuation={ytConfigData.Continuation}";
-            }
-
-            string jsonContent = GetRequestPayloadData(ytConfigData);
-
-            HttpRequestMessage httpRequestMessage = new(HttpMethod.Post, url);
-
-            if (!string.IsNullOrEmpty(SharedCookies))
-            {
-                SetHttpRequestMessageHeader(httpRequestMessage, ytConfigData);
-            }
-
-            HttpContent httpContent = new StringContent(jsonContent, Encoding.UTF8, "application/json");
-
-            httpRequestMessage.Content = httpContent;
-
-            HttpResponseMessage? httpResponseMessage = SharedHttpClient?.SendAsync(httpRequestMessage)
-                .GetAwaiter()
-                .GetResult();
-
-            string? receivedJsonContent = httpResponseMessage?.Content.ReadAsStringAsync()
-                .GetAwaiter()
-                .GetResult();
-
-            RaiseOnLogOutput(
-                EnumSet.LogType.Debug,
-                httpRequestMessage.ToString());
-            RaiseOnLogOutput(
-                EnumSet.LogType.Debug,
-                jsonContent);
-            RaiseOnLogOutput(
-                EnumSet.LogType.Debug,
-                httpResponseMessage?.ToString() ?? string.Empty);
-            RaiseOnLogOutput(
-                EnumSet.LogType.Debug,
-                receivedJsonContent ?? string.Empty);
-
-            if (string.IsNullOrEmpty(receivedJsonContent))
-            {
-                RaiseOnLogOutput(
-                    EnumSet.LogType.Error,
-                    "[GetJsonElement()] 發生錯誤，變數 \"receivedJsonContent\" 為空白或是 null！");
-
-                return jsonElement;
-            }
-
-            if (httpResponseMessage?.StatusCode == HttpStatusCode.OK)
-            {
-                jsonElement = JsonSerializer.Deserialize<JsonElement>(receivedJsonContent);
-            }
-            else
-            {
-                string errorMessage = $"[{DateTime.Now}]（{typeof(HttpClient).Name}）：" +
-                    $"連線發生錯誤，錯誤碼：{httpResponseMessage?.StatusCode} " +
-                    $"{(httpResponseMessage != null ? $"({(int)(httpResponseMessage.StatusCode)})" : string.Empty)}{Environment.NewLine}" +
-                    $"接收到的內容：{Environment.NewLine}" +
-                    $"{receivedJsonContent}{Environment.NewLine}";
-
-                RaiseOnLogOutput(
-                    EnumSet.LogType.Error,
-                    errorMessage);
-            }
-        }
-
-        return jsonElement;
-    }
-
-    /// <summary>
-    /// 取得要求裝載資料
-    /// </summary>
-    /// <param name="ytConfigData">YTConfigData</param>
-    /// <returns>字串</returns>
-    private string GetRequestPayloadData(YTConfigData ytConfigData)
-    {
-        // 參考：https://github.com/xenova/chat-downloader/blob/master/chat_downloader/sites/youtube.py#L1764
-        // 參考：https://github.com/abhinavxd/youtube-live-chat-downloader/blob/main/yt_chat.go
-
-        // 內容是精簡過的。
-        RequestPayloadData requestPayloadData = new()
-        {
-            Context = new()
-            {
-                Client = new()
-                {
-                    BrowserName = ytConfigData.BrowserName ?? string.Empty,
-                    BrowserVersion = ytConfigData.BrowserVersion ?? string.Empty,
-                    ClientFormFactor = ytConfigData.ClientFormFactor ?? "UNKNOWN_FORM_FACTOR",
-                    ClientName = ytConfigData.ClientName,
-                    ClientVersion = ytConfigData.ClientVersion,
-                    DeviceMake = ytConfigData.DeviceMake ?? string.Empty,
-                    DeviceModel = ytConfigData.DeviceModel ?? string.Empty,
-                    // 語系會影響取得的內容，使用 zh-TW, TW。
-                    Gl = ytConfigData.Gl ?? "TW",
-                    Hl = ytConfigData.Hl ?? "zh-TW",
-                    OriginalUrl = ytConfigData.OriginalUrl ?? string.Empty,
-                    OsName = ytConfigData.OsName ?? "Windows",
-                    OsVersion = ytConfigData.OsVersion ?? "10.0",
-                    Platform = ytConfigData.Platform ?? "DESKTOP",
-                    RemoteHost = ytConfigData.RemoteHost ?? "DESKTOP",
-                    UserAgent = ytConfigData.UserAgent ?? string.Empty,
-                    VisitorData = ytConfigData.VisitorData,
-                    TimeZone = "Asia/Taipei"
-                }
-            },
-            Continuation = ytConfigData.Continuation
-        };
-
-        #region 最後總檢查 gl、hl 的值
-
-        if (requestPayloadData.Context.Client.Gl != "TW")
-        {
-            requestPayloadData.Context.Client.Gl = "TW";
-        }
-
-        if (requestPayloadData.Context.Client.Hl != "zh-TW")
-        {
-            requestPayloadData.Context.Client.Hl = "zh-TW";
-        }
-
-        #endregion
-
-        return JsonSerializer.Serialize(requestPayloadData);
-    }
-
-    #region YouTube 會員驗證機制
-
-    /// <summary>
-    /// 設定 HttpRequestMessage 的標頭
-    /// <para>來源：https://stackoverflow.com/a/13287224</para>
-    /// <para>原作者：Greg Beech</para>
-    /// <para>原授權：CC BY-SA 3.0</para>
-    /// <para>CC BY-SA 3.0：https://creativecommons.org/licenses/by-sa/3.0/</para>
-    /// </summary>
-    /// <param name="httpRequestMessage">HttpRequestMessage</param>
-    /// <param name="ytConfigData">YTConfigData</param>
-    private void SetHttpRequestMessageHeader(
-        HttpRequestMessage httpRequestMessage,
-        YTConfigData? ytConfigData = null)
-    {
-        if (!string.IsNullOrEmpty(SharedCookies))
-        {
-            httpRequestMessage.Headers.Add("Cookie", SharedCookies);
-
-            string[] SharedCookiesArray = SharedCookies.Split(
-                ';',
-                StringSplitOptions.RemoveEmptyEntries);
-
-            string? sapiSid = SharedCookiesArray.FirstOrDefault(n => n.Contains("SAPISID"));
-
-            if (!string.IsNullOrEmpty(sapiSid))
-            {
-                string[] tempArray = sapiSid.Split(
-                    '=',
-                    StringSplitOptions.RemoveEmptyEntries);
-
-                if (tempArray.Length == 2)
-                {
-                    httpRequestMessage.Headers.Authorization =
-                        new AuthenticationHeaderValue(
-                            "SAPISIDHASH",
-                            GetSapiSidHash(tempArray[1], StringSet.Origin));
-                }
-            }
-        }
-
-        if (ytConfigData != null)
-        {
-            string xGoogAuthuser = "0",
-                xGoogPageId = string.Empty;
-
-            if (!string.IsNullOrEmpty(ytConfigData.DataSyncID))
-            {
-                xGoogPageId = ytConfigData.DataSyncID;
-            }
-
-            if (string.IsNullOrEmpty(xGoogPageId) &&
-                !string.IsNullOrEmpty(ytConfigData.DelegatedSessionID))
-            {
-                xGoogPageId = ytConfigData.DelegatedSessionID;
-            }
-
-            if (!string.IsNullOrEmpty(xGoogPageId))
-            {
-                httpRequestMessage.Headers.Add("X-Goog-Pageid", xGoogPageId);
-            }
-
-            if (!string.IsNullOrEmpty(ytConfigData.IDToken))
-            {
-                httpRequestMessage.Headers.Add("X-Youtube-Identity-Token", ytConfigData.IDToken);
-            }
-
-            if (!string.IsNullOrEmpty(ytConfigData.SessionIndex))
-            {
-                xGoogAuthuser = ytConfigData.SessionIndex;
-            }
-
-            httpRequestMessage.Headers.Add("X-Goog-Authuser", xGoogAuthuser);
-            httpRequestMessage.Headers.Add("X-Goog-Visitor-Id", ytConfigData.VisitorData);
-            httpRequestMessage.Headers.Add("X-Youtube-Client-Name", ytConfigData.InnetrubeContextClientName.ToString());
-            httpRequestMessage.Headers.Add("X-Youtube-Client-Version", ytConfigData.InnetrubeClientVersion);
-
-            if (!string.IsNullOrEmpty(ytConfigData.InitPage))
-            {
-                httpRequestMessage.Headers.Referrer = new Uri(ytConfigData.InitPage);
-            }
-        }
-
-        httpRequestMessage.Headers.Add("Origin", StringSet.Origin);
-        httpRequestMessage.Headers.Add("X-Origin", StringSet.Origin);
-    }
-
-    /// <summary>
-    /// 取得 SAPISIDHASH 字串
-    /// </summary>
-    /// <param name="sapiSid">字串，SAPISID</param>
-    /// <param name="origin">字串，origin</param>
-    /// <returns>字串</returns>
-    private string GetSapiSidHash(string sapiSid, string origin)
-    {
-        long unixTimestamp = DateTimeOffset.Now.ToUnixTimeSeconds();
-
-        return $"{unixTimestamp}_{GetSHA1Hash($"{unixTimestamp} {sapiSid} {origin}")}";
-    }
-
-    /// <summary>
-    /// 取得 SHA-1 雜湊 
-    /// </summary>
-    /// <param name="value">字串，值</param>
-    /// <returns>字串</returns>
-    private string GetSHA1Hash(string value)
-    {
-        byte[] bytes = SHA1.HashData(Encoding.UTF8.GetBytes(value));
-
-        StringBuilder builder = new();
-
-        for (int i = 0; i < bytes.Length; i++)
-        {
-            builder.Append(bytes[i].ToString("x2"));
-        }
-
-        return builder.ToString();
-    }
-
-    #endregion
-
-    #region 解析 JSON 資料
-
     /// <summary>
     /// 解析 ytcfg 
     /// </summary>
@@ -1940,47 +1534,4 @@ public partial class LiveChatCatcher
 
         return output;
     }
-
-    #endregion
-
-    #region 公用方法
-
-    /// <summary>
-    /// 取得 Hex 色碼
-    /// </summary>
-    /// <param name="value">Int64</param>
-    /// <returns>字串</returns>
-    private string GetColorHexCode(long value)
-    {
-        string hex = string.Format("{0:X}", value);
-
-        int integer = Convert.ToInt32(hex, 16);
-
-        return ColorTranslator.ToHtml(Color.FromArgb(integer));
-    }
-
-    /// <summary>
-    /// 取得 RendererData 的 Type
-    /// </summary>
-    /// <param name="rendererName">字串，*Renderer 的名稱</param>
-    /// <returns>字串</returns>
-    private string GetRendererDataType(string rendererName)
-    {
-        return rendererName switch
-        {
-            "liveChatTextMessageRenderer" => StringSet.ChatGeneral,
-            "liveChatPaidMessageRenderer" => StringSet.ChatSuperChat,
-            "liveChatPaidStickerRenderer" => StringSet.ChatSuperSticker,
-            "liveChatMembershipItemRenderer" => StringSet.ChatJoinMember,
-            "liveChatViewerEngagementMessageRenderer" => StringSet.YouTube,
-            "liveChatModeChangeMessageRenderer" => StringSet.YouTube,
-            "liveChatSponsorshipsGiftPurchaseAnnouncementRenderer" => StringSet.ChatMemberGift,
-            "liveChatSponsorshipsGiftRedemptionAnnouncementRenderer" => StringSet.ChatReceivedMemberGift,
-            "liveChatBannerHeaderRenderer" => StringSet.ChatPinned,
-            "liveChatBannerRedirectRenderer" => StringSet.ChatRedirect,
-            _ => string.Empty
-        };
-    }
-
-    #endregion 公用方法
 }
