+++ README.md
@@ -2,7 +2,7 @@
 
 ## 簡述
 
-模擬 YouTube 直播聊天室網頁版的行為，以取得直播中或重播時的聊天室內容，並提供簡易的收益分析（新臺幣）、報表匯出等等功能。
+模擬 YouTube 直播聊天室網頁版的行為，以取得直播中或一般影片的聊天室內容，並提供簡易的收益分析（新臺幣）、報表匯出等等功能。
 
 ## 請先讀我
